Fix code review issues
- Replace Spring @Nullable with jakarta.annotation.Nullable in ImportTaskMessage
- Fix getTime() to use Long.parseLong instead of Integer.parseInt in CollectRep
- Add jakarta.annotation-api dependency to common-core

+++ hertzbeat-common-core/src/main/java/org/apache/hertzbeat/common/entity/message/CollectRep.java
@@ -486,7 +486,7 @@ public final class CollectRep {
             }
 
             public long getTime() {
-                return Integer.parseInt(metadata.getOrDefault(MetricDataConstants.TIME, "0"));
+                return Long.parseLong(metadata.getOrDefault(MetricDataConstants.TIME, "0"));
             }
 
             public Code getCode() {
